[MOV] figure: move figure menu items in another file
This commit moves the figure menu items to a new file to avoid
circular dependencies in the next commit.

Task: 4497382
